+++ packages/datadog_grpc_interceptor/lib/src/datadog_grpc_interceptor.dart
@@ -70,9 +70,9 @@ class DatadogGrpcInterceptor extends ClientInterceptor {
             rum.tracingSamplingRate / 100.0;
         if (tracingContext.sampled) {
           attributes[DatadogRumPlatformAttributeKey.traceID] =
-              tracingContext.traceId!.asString(TraceIdRepresentation.decimal);
+              tracingContext.traceId.asString(TraceIdRepresentation.decimal);
           attributes[DatadogRumPlatformAttributeKey.spanID] =
-              tracingContext.spanId!.asString(TraceIdRepresentation.decimal);
+              tracingContext.spanId.asString(TraceIdRepresentation.decimal);
         }
 
         for (final tracingType in tracingHeaderTypes) {