+++ packages/datadog_grpc_interceptor/CHANGELOG.md
@@ -3,7 +3,7 @@
 ## Unreleased
 
 * Send trace sample rate (`_dd.rule_psr`) for APM's traffic ingestion control page.
-* Support for OTel b3 header injection
+* Support for OTel `b3` and W3C `tracecontext` header injection
 
 ## 1.0.0-beta.1