+++ packages/datadog_tracking_http_client/CHANGELOG.md
@@ -6,7 +6,7 @@
 * Send trace sample rate (`dd.rule_psr`) for APM's traffic ingestion control page.
 * Added DatadogClient for use with the `http` pub package.
 * Fix an issue where convenience methods on DataodgTrackingHttpClient weren't being tracked properly
-* Support for OTel b3 header injection
+* Support for OTel `b3` and W3C `tracecontext` header injection
 
 ## 1.1.0